Refactored commands
